requirements.txt
requests==2.31.0
cryptography==41.0.7
baidu-aip==2.2.18.0